requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
diff-match-patch==20241021
Django==4.2.17
django-admin-autocomplete-filter==0.7.1
django-environ==0.11.2
django-import-export==4.1.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==10.4.0
sqlparse==0.5.3
tablib==3.5.0
typing_extensions==4.12.2
